Fix the VMS build for my C++ fork of bash.
Fixed my C++ fork of bash to compile with OpenVMS. Now I'm working to
refactor it so that it works with vfork()/execve() instead of fork().

+++ reference/bash/execute_cmd.c
@@ -3070,7 +3070,7 @@ displen (const char *s)
 
   wcstr = 0;
   slen = mbstowcs (wcstr, s, 0);
-  if (slen == -1)
+  if (slen == (size_t)-1)
     slen = 0;
   wcstr = (wchar_t *)xmalloc (sizeof (wchar_t) * (slen + 1));
   mbstowcs (wcstr, s, slen + 1);
